feat(utils): move charCount, lineCount and wordCound into utils
