--- packages/counter/.babelrc
@@ -1,10 +1,0 @@
-{
-  "presets": [
-  ["@babel/preset-env",  {
-    "targets": "> 0.25%, not dead"
-  }],
-  "@babel/preset-flow",
-  "@babel/preset-react"
-  ],
-  "plugins": ["@babel/plugin-proposal-class-properties"]
-}
