--- packages/counter/.flowconfig
@@ -1,17 +1,0 @@
-[ignore]
-.*/node_modules/.*
-
-[include]
-node_modules/draft-js
-node_modules/react
-
-[libs]
-
-[lints]
-
-[options]
-
-[strict]
-
-[version]
-^0.81.0
