--- packages/counter/rollup.config.js
@@ -1,38 +1,0 @@
-import babel from 'rollup-plugin-babel'
-import commonjs from 'rollup-plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import postcss from 'rollup-plugin-postcss'
-import resolve from 'rollup-plugin-node-resolve'
-import url from 'rollup-plugin-url'
-
-import pkg from './package.json'
-
-export default {
-  external: ['@djsp/editor', 'draft-js'],
-  input: 'src/index.js',
-  output: [
-    {
-      file: pkg.main,
-      format: 'cjs',
-    },
-    {
-      file: pkg.module,
-      format: 'es',
-    },
-  ],
-  plugins: [
-    external(),
-    postcss({
-      modules: true,
-      extract: true,
-    }),
-    url(),
-    babel({
-      exclude: 'node_modules/**',
-    }),
-    resolve({
-      preferBuiltins: false,
-    }),
-    commonjs(),
-  ],
-}
