--- packages/counter/src/utils/getLineCount.js
@@ -1,4 +1,0 @@
-export default function getLineCount(editorState) {
-  const blockArray = editorState.getCurrentContent().getBlocksAsArray()
-  return blockArray ? blockArray.length : null
-}
