+++ packages/utils/package.json
@@ -25,6 +25,7 @@
     "cross-env": "^5.1.4",
     "flow-bin": "^0.81.0",
     "flow-typed": "^2.5.1",
+    "punycode": "^2.1.1",
     "react": "^16.2.0",
     "react-dom": "^16.6.0",
     "rollup": "^0.66.2",
